Moved Hierarchy-related stuff to a new Hierarchy.ts file
