+++ static/components/hierarchy-node.html
@@ -1,8 +1,8 @@
 
 <li>
-    <div class="hierarchyNode" data-bind="class: $root.isHierarchyNodeSelected($root.objectIsSelectedById($data._id)), css:{'hierarchyCollapsed':$data.expanded()===false}, attr: {id:key} , click:function(node, e){$root.selectNodeInHierarchy($data, e);}, clickBubble:false">
+    <div class="hierarchyNode" data-bind="class: Hierarchy.isNodeSelected($root.objectIsSelectedById($data._id)), css:{'hierarchyCollapsed':$data.expanded()===false}, attr: {id:key} , click:function(node, e){Hierarchy.selectNode($data, e);}, clickBubble:false">
         <div class="hierarchyNodeColor" data-bind=" style: {'background-color': color}"></div>
-        <div class="hierarchyNodeVisibilityIcon" data-bind="style: {'visibility': $root.hierarchyNodeIsHidden($data.key())}">
+        <div class="hierarchyNodeVisibilityIcon" data-bind="style: {'visibility': Hierarchy.nodeIsHidden($data.key())}">
             <i class="material-icons md-18">visibility_off</i>
         </div>
             <!-- ko if: isGroup() -->
@@ -14,7 +14,7 @@
         <span data-bind="text: category"></span>
         <div>
             <!-- ko with: inputApplication -->
-                <div class="hierarchyNode hierarchyApplication" data-bind="class: $root.isHierarchyApplicationSelected($root.objectIsSelectedById($data._id)),attr:{id:key}, style:{'background-color':$parent.color}, click:function(node, e){$root.selectNodeInHierarchy($data, e);}, clickBubble:false">
+                <div class="hierarchyNode hierarchyApplication" data-bind="class: Hierarchy.isApplicationSelected($root.objectIsSelectedById($data._id)),attr:{id:key}, style:{'background-color':$parent.color}, click:function(node, e){Hierarchy.selectNode($data, e);}, clickBubble:false">
                     <div class="hierarchyNodeColor" data-bind=" style: {'background-color': color}"></div>
                     <h5 data-bind="text: name"></h5>
                     <span data-bind="text: category"></span>
@@ -26,7 +26,7 @@
             </ul>
 
             <!-- ko with: outputApplication -->
-                <div class="hierarchyNode hierarchyApplication" data-bind="class: $root.isHierarchyApplicationSelected($root.objectIsSelectedById($data._id)), attr:{id:key}, style:{'background-color':$parent.color}, click:function(node, e){$root.selectNodeInHierarchy($data, e);}, clickBubble:false">
+                <div class="hierarchyNode hierarchyApplication" data-bind="class: Hierarchy.isApplicationSelected($root.objectIsSelectedById($data._id)), attr:{id:key}, style:{'background-color':$parent.color}, click:function(node, e){Hierarchy.selectNode($data, e);}, clickBubble:false">
                     <div class="hierarchyNodeColor" data-bind=" style: {'background-color': color}"></div>
                     <h5 data-bind="text: name"></h5>
                     <span data-bind="text: category"></span>
