+++ tsconfig.json
@@ -46,6 +46,7 @@
         "./src/PaletteInfo.ts",
         "./src/ExplorePalettes.ts",
         "./src/Undo.ts",
-        "./src/Errors.ts"
+        "./src/Errors.ts",
+        "./src/Hierarchy.ts"
     ]
 }
